refactor: IdeaMarketEdit.tsx 코드 리팩토링
