+++ src/components/my-page/edit/imageUploader.module.scss
@@ -2,25 +2,6 @@
   margin-bottom: 1.5rem;
 }
 
-.imageUploadButton {
-  width: 100px;
-  height: 100px;
-  border: 2px dashed $grey-300;
-  border-radius: 4px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  background: none;
-}
-
-.imageIcon {
-  font-size: 24px;
-  color: $grey-500;
-  margin-bottom: 0.5rem;
-}
-
 .imagePreviewContainer {
   width: 100px;
   height: 100px;
@@ -36,3 +17,31 @@
   object-fit: contain;
   background-color: white;
 }
+
+.mainImage {
+  width: 100px;
+  height: 100px;
+  cursor: pointer;
+}
+
+// .imageUploadButton {
+//   width: 100px;
+//   height: 100px;
+//   border: 2px dashed $grey-300;
+//   border-radius: 4px;
+//   display: flex;
+//   flex-direction: column;
+//   align-items: center;
+//   justify-content: center;
+//   cursor: pointer;
+//   background: none;
+// }
+
+// .imageIcon {
+//   font-size: 24px;
+//   color: $grey-500;
+//   margin-bottom: 0.5rem;
+// }
+
+
+
