+++ src/components/my-page/edit/pageSelect.module.scss
@@ -2,13 +2,53 @@
   margin-bottom: 1.5rem;
 }
 
-.categoryGroup {
-  width: 380px;
-  margin-bottom: 0;
-}
-
 .pageTypeGroup {
   position: relative;
+}
+
+.pageTypeWrapper {
+  display: flex;
+  width: 380px;
+  height: 51px;
+  align-items: center;
+  border: 1px solid $grey-300;
+  border-radius: 5px;
+  background: white;
+  overflow: hidden;
+}
+
+.pageTypeButton {
+  width: 50%;
+  height: 100%;
+  padding: 0;
+  border: none;
+  background: transparent;
+  cursor: pointer;
+  position: relative;
+  
+  color: $grey-600;
+  text-align: center;
+  font-family: $font-family;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  letter-spacing: -0.18px;
+
+  &:first-child::after {
+    content: '';
+    position: absolute;
+    right: 0;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 1px;
+    height: 16px;
+    background-color: $grey-300;
+  }
+
+  &.active {
+    color: $primary;
+  }
 }
 
 .pageDescription {
@@ -74,4 +114,22 @@
       letter-spacing: -0.12px;
     }
   }
+}
+
+.labelWrapper {
+  margin-bottom: 20px;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  color: $grey-800;
+  font-family: $font-family;
+  font-size: 20px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  letter-spacing: -0.2px;
+}
+
+.required {
+  color: $red;
 }
